requirements.txt
streamlit>=1.26.0
tensorflow==2.12.1
protobuf>=3.20.3
Pillow
numpy
